Major Updates
Filled out majority of content, including the full about and support sections. Also, started working on the layout and content for both the archives, and the store. Additionally, updated jumbotron with most recent plot photo. Made lots of small changes as well, but (a) can't remember all of them and (b) there would be too many to list here anyways.